requirements.txt
# Gemini Live Hebrew STS - Python Dependencies

# Google AI SDK for Gemini
google-genai>=0.3.0

# Audio processing
pyaudio>=0.2.14

# WebSocket support
websockets>=12.0

# Async support (usually built-in, but explicit for clarity)
aiohttp>=3.9.0

# Raspberry Pi GPIO (optional, only for hardware integration)
# Install on Raspberry Pi with: sudo apt-get install python3-rpi.gpio
# RPi.GPIO>=0.7.1

# Optional: for environment variable management
python-dotenv>=1.0.0
